Extract rotation strategy module
Split FileRotationStrategy into rust_extension/src/handlers/rotating/strategy.rs so the rotating handler module drops below the 400-line limit while keeping the existing API surface available via a pub(crate) re-export.
